+++ pythonFiles/include/blender_vscode/communication.py
@@ -1,4 +1,3 @@
-import bpy
 import time
 import flask
 import debugpy
@@ -6,12 +5,13 @@ import random
 import requests
 import threading
 from functools import partial
-from . utils import run_in_main_thread
-from . environment import blender_path, scripts_folder
+from .utils import run_in_main_thread
+from .environment import blender_path, scripts_folder
 
 EDITOR_ADDRESS = None
 OWN_SERVER_PORT = None
 DEBUGPY_PORT = None
+
 
 def setup(address, path_mappings):
     global EDITOR_ADDRESS, OWN_SERVER_PORT, DEBUGPY_PORT
@@ -25,6 +25,7 @@ def setup(address, path_mappings):
     print("Waiting for debug client.")
     debugpy.wait_for_client()
     print("Debug client attached.")
+
 
 def start_own_server():
     port = [None]
@@ -46,6 +47,7 @@ def start_own_server():
 
     return port[0]
 
+
 def start_debug_server():
     while True:
         port = get_random_port()
@@ -63,7 +65,8 @@ def start_debug_server():
 server = flask.Flask("Blender Server")
 post_handlers = {}
 
-@server.route("/", methods=['POST'])
+
+@server.route("/", methods=["POST"])
 def handle_post():
     data = flask.request.get_json()
     print("Got POST:", data)
@@ -73,7 +76,8 @@ def handle_post():
 
     return "OK"
 
-@server.route("/", methods=['GET'])
+
+@server.route("/", methods=["GET"])
 def handle_get():
     flask.request
     data = flask.request.get_json()
@@ -88,25 +92,31 @@ def register_post_handler(type, handler):
     assert type not in post_handlers
     post_handlers[type] = handler
 
+
 def register_post_action(type, handler):
     def request_handler_wrapper(data):
         run_in_main_thread(partial(handler, data))
         return "OK"
+
     register_post_handler(type, request_handler_wrapper)
 
 
 # Sending Data
 ###############################
 
+
 def send_connection_information(path_mappings):
-    send_dict_as_json({
-        "type" : "setup",
-        "blenderPort" : OWN_SERVER_PORT,
-        "debugpyPort" : DEBUGPY_PORT,
-        "blenderPath" : str(blender_path),
-        "scriptsFolder" : str(scripts_folder),
-        "addonPathMappings" : path_mappings,
-    })
+    send_dict_as_json(
+        {
+            "type": "setup",
+            "blenderPort": OWN_SERVER_PORT,
+            "debugpyPort": DEBUGPY_PORT,
+            "blenderPath": str(blender_path),
+            "scriptsFolder": str(scripts_folder),
+            "addonPathMappings": path_mappings,
+        }
+    )
+
 
 def send_dict_as_json(data):
     print("Sending:", data)
@@ -116,14 +126,18 @@ def send_dict_as_json(data):
 # Utils
 ###############################
 
+
 def get_random_port():
     return random.randint(2000, 10000)
+
 
 def get_blender_port():
     return OWN_SERVER_PORT
 
+
 def get_debugpy_port():
     return DEBUGPY_PORT
+
 
 def get_editor_address():
     return EDITOR_ADDRESS