+++ pythonFiles/include/blender_vscode/operators/addon_update.py
@@ -1,9 +1,11 @@
+from pathlib import Path
 import bpy
 import sys
 import traceback
 from bpy.props import *
-from .. utils import redraw_all
-from .. communication import send_dict_as_json, register_post_action
+from ..utils import is_addon_legacy, redraw_all
+from ..communication import send_dict_as_json, register_post_action
+
 
 class UpdateAddonOperator(bpy.types.Operator):
     bl_idname = "dev.update_addon"
@@ -14,10 +16,10 @@ class UpdateAddonOperator(bpy.types.Operator):
     def execute(self, context):
         try:
             bpy.ops.preferences.addon_disable(module=self.module_name)
-        except:
+        except Exception:
             traceback.print_exc()
-            send_dict_as_json({"type" : "disableFailure"})
-            return {'CANCELLED'}
+            send_dict_as_json({"type": "disableFailure"})
+            return {"CANCELLED"}
 
         for name in list(sys.modules.keys()):
             if name.startswith(self.module_name):
@@ -25,19 +27,28 @@ class UpdateAddonOperator(bpy.types.Operator):
 
         try:
             bpy.ops.preferences.addon_enable(module=self.module_name)
-        except:
+        except Exception:
             traceback.print_exc()
-            send_dict_as_json({"type" : "enableFailure"})
-            return {'CANCELLED'}
+            send_dict_as_json({"type": "enableFailure"})
+            return {"CANCELLED"}
 
-        send_dict_as_json({"type" : "addonUpdated"})
+        send_dict_as_json({"type": "addonUpdated"})
 
         redraw_all()
-        return {'FINISHED'}
+        return {"FINISHED"}
+
 
 def reload_addon_action(data):
-    for name in data["names"]:
+    module_names = []
+    for name, dir in zip(data["names"], data["dirs"]):
+        if is_addon_legacy(Path(dir)):
+            module_names.append(name)
+        else:
+            module_names.append("bl_ext.user_default." + name)
+
+    for name in module_names:
         bpy.ops.dev.update_addon(module_name=name)
+
 
 def register():
     bpy.utils.register_class(UpdateAddonOperator)