+++ pythonFiles/include/blender_vscode/operators/script_runner.py
@@ -2,8 +2,9 @@ import re
 import bpy
 import runpy
 from bpy.props import *
-from .. utils import redraw_all
-from .. communication import register_post_action
+from ..utils import redraw_all
+from ..communication import register_post_action
+
 
 class RunScriptOperator(bpy.types.Operator):
     bl_idname = "dev.run_script"
@@ -13,22 +14,29 @@ class RunScriptOperator(bpy.types.Operator):
 
     def execute(self, context):
         ctx = prepare_script_context(self.filepath)
-        runpy.run_path(self.filepath, init_globals={"CTX" : ctx})
+        runpy.run_path(self.filepath, init_globals={"CTX": ctx})
         redraw_all()
-        return {'FINISHED'}
+        return {"FINISHED"}
 
 
 def run_script_action(data):
     path = data["path"]
     context = prepare_script_context(path)
-    bpy.ops.dev.run_script(context, filepath=path)
+
+    if bpy.app.version < (4, 0, 0):
+        bpy.ops.dev.run_script(context, filepath=path)
+        return
+
+    with bpy.context.temp_override(**context):
+        bpy.ops.dev.run_script(filepath=path)
+
 
 def prepare_script_context(filepath):
     with open(filepath) as fs:
         text = fs.read()
 
-    area_type = 'VIEW_3D'
-    region_type = 'WINDOW'
+    area_type = "VIEW_3D"
+    region_type = "WINDOW"
 
     for line in text.splitlines():
         match = re.match(r"^\s*#\s*context\.area\s*:\s*(\w+)", line, re.IGNORECASE)
@@ -42,17 +50,10 @@ def prepare_script_context(filepath):
     context["view_layer"] = context["window"].view_layer
     context["screen"] = context["window"].screen
     context["workspace"] = context["window"].workspace
-    context["active_object"] = context["view_layer"].objects.active
-    context["object"] = context["active_object"]
-    context["collection"] = context["view_layer"].active_layer_collection.collection
     context["area"] = get_area_by_type(area_type)
-    context["space_data"] = context["area"].spaces.active
-    context["selected_objects"] = [obj for obj in context["view_layer"].objects if obj.select_get(view_layer=context["view_layer"])]
-    context["selected_editable_objects"] = context["selected_objects"]
     context["region"] = get_region_in_area(context["area"], region_type) if context["area"] else None
-    if context["space_data"].type == "VIEW_3D":
-        context["region_data"] = context["space_data"].region_3d
     return context
+
 
 def get_area_by_type(area_type):
     for area in bpy.data.window_managers[0].windows[0].screen.areas:
@@ -60,11 +61,13 @@ def get_area_by_type(area_type):
             return area
     return None
 
+
 def get_region_in_area(area, region_type):
     for region in area.regions:
         if region.type == region_type:
             return region
     return None
+
 
 def register():
     bpy.utils.register_class(RunScriptOperator)