+++ pythonFiles/templates/script.py
@@ -1,5 +1,5 @@
 import bpy
 from mathutils import *
+
 D = bpy.data
 C = bpy.context
-