+++ src/addon_folder.ts
@@ -2,8 +2,11 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import {
     getConfig, readTextFile, getWorkspaceFolders,
-    getSubfolders, executeTask
+    getSubfolders, executeTask, getAnyWorkspaceFolder, pathExists
 } from './utils';
+
+// TODO: It would be superior to use custom AddonFolder interface that is not bound to the
+// vscode.WorkspaceFolder directly. The 'uri' property is only one used at this point.
 
 export class AddonWorkspaceFolder {
     folder: vscode.WorkspaceFolder;
@@ -13,8 +16,16 @@ export class AddonWorkspaceFolder {
     }
 
     public static async All() {
+        // Search folders specified by settings first, if nothing is specified
+        // search workspace folders instead.
+        let addonFolders = await foldersToWorkspaceFoldersMockup(
+            <string[]>getConfig().get('addonFolders'));
+        if (addonFolders.length === 0) {
+            addonFolders = getWorkspaceFolders();
+        }
+
         let folders = [];
-        for (let folder of getWorkspaceFolders()) {
+        for (let folder of addonFolders) {
             let addon = new AddonWorkspaceFolder(folder);
             if (await addon.hasAddonEntryPoint()) {
                 folders.push(addon);
@@ -33,6 +44,10 @@ export class AddonWorkspaceFolder {
 
     get reloadOnSave() {
         return <boolean>this.getConfig().get('addon.reloadOnSave');
+    }
+
+    get justMyCode() {
+        return <boolean>this.getConfig().get('addon.justMyCode');
     }
 
     public async hasAddonEntryPoint() {
@@ -115,6 +130,11 @@ async function tryFindActualAddonFolder(root: string) {
 }
 
 async function folderContainsAddonEntry(folderPath: string) {
+    let manifestPath = path.join(folderPath, "blender_manifest.toml");
+    if (await pathExists(manifestPath)) {
+        return true;
+    }
+        
     let initPath = path.join(folderPath, '__init__.py');
     try {
         let content = await readTextFile(initPath);
@@ -123,4 +143,32 @@ async function folderContainsAddonEntry(folderPath: string) {
     catch {
         return false;
     }
+}
+
+async function foldersToWorkspaceFoldersMockup(folders: string[]) {
+    let mockups: vscode.WorkspaceFolder[] = [];
+    // Assume this functionality is only used with a single workspace folder for now.
+    let rootFolder = getAnyWorkspaceFolder();
+    for (let i = 0; i < folders.length; i++) {
+        let absolutePath;
+        if (path.isAbsolute(folders[i])) {
+            absolutePath = folders[i];
+        } else {
+            absolutePath = path.join(rootFolder.uri.fsPath, folders[i])
+        }
+
+        let exists = await pathExists(absolutePath);
+        if (!exists) {
+            vscode.window.showInformationMessage(
+                `Revise settings, path to addon doesn't exist ${absolutePath}`);
+            continue;
+        }
+
+        mockups.push({
+            "name" : path.basename(absolutePath),
+            "uri": vscode.Uri.parse(absolutePath),
+            "index": i
+        });
+    }
+    return mockups;
 }